Support advanced "vm config" options for topology nodes
For example, in the sample topology spec below, we add an advanced
option for the `qemu-append` minimega vm config option:

```
spec:
  nodes:
  - type: VirtualMachine
    general:
      hostname: host-01
      snapshot: true
    hardware:
      os_type: linux
      drives:
      - image: miniccc.qc2
    advanced:
      qemu-append: -rtc driftfix=slew -cpu Haswell-noTSX-IBRS,hv_relaxed,hv_spinlocks=0x1fff,hv_vapic,hv_time
```

Note that by default phenix sets qemu-append to `-vga qxl` if the VM is
running Linux. In the above example, `-vga qxl` would be replaced by
what's in the node config.

Also note that **any** minimega vm config option can be specified in the
`advanced` section, and will replace the config option if it was already
set previously. The advanced options get set at the very end of the vm
config step. As an example of this, the above example sets the minimega
vm config snapshot option to true. If one were to specify `snapshot:
"false"` in the `advanced` section of the node config, it would
overwrite the snapshot setting specified in the `general` section.

closes #74