Accurately display VM disk from topology, regardless of location on disk
VM disk image paths defined in a topology can either be relative or
absolute paths. If relative, they're assumed to be relative from the
configured images directory (which by default is `/phenix/images`).

This commit updates the UI to always show the full filename (including
extension) only of the VM disk image, regardless of whether it was
defined as a relative or absolute path in the topology file. This
assumes the VM disk image referenced actually exists on disk. If not,
the original behavior of showing nothing in the dropdown still occurs.

This commit also adds a tooltip to the disk dropdown to show the full
path of the disk image currently selected. If more than one disk image
of the same name is found on disk, preference is given to images found
in the configured images directory (e.g., `/phenix/images`) for the
tooltip.

closes #49