Enable using minimega C2 on Windows 10 VMs to run startup script
This obviously requires that the Windows 10 VM image be configured ahead
of time to run the miniccc agent at boot. Since the `image inject-miniexe`
subcommand uses the startup folder script technique to start the miniccc
agent, the miniccc agent and service must be added to Windows 10 images
in a different manner (e.g. a Packer config).