Add support for custom reachability tests in SoH app
Support configuration of TCP and UDP reachability tests in addition to
the ICMP-based reachability testing SoH already supports.

For UDP tests, a valid UDP payload that will trigger a UDP response must
be provided as a Base64-encoded string.

Note that this commit relies on new features recently added to minimega
that haven't been merged into the default minimega branch yet. The
minimega code required is in the activeshadow/minimega@tcp-conn-test
branch.