Add banner to VM VNC windows
If a VM has an annotation named `vncBanner` in its topology config, then
a banner will be added to the top and bottom of the VNC tab for the VM.

If the value of the annotation is a simple string, then the string will
be used for the top and bottom banner, and the banner will have a white
background with black text.

If the value of the annotation is a map, it should have the following
schema:

```
topBanner:
  banner:
  - Banner Text Line 1
  - Banner Text Line 2
  - Banner Text Line n
  backgroundColor: red
  textColor: yellow
bottomBanner:
  banner:
  - Banner Text Line 1
  - Banner Text Line 2
  - Banner Text Line n
  backgroundColor: red
  textColor: yellow
```

If the above schema is used, the top or bottom banner can be omitted.
Values for the background and text color settings can be any valid HTML
color values (e.g., names of colors or color codes).