Add topology builder to UI
Integrate topology builder from minimega (minibuilder), adding support
for saving topologies and experiments directly to the phenix store.

When saving a builder drawing as a topology in phenix, a `builder-xml`
annotation is added to the topology config. This XML is then used to
load the topology back into builder if/when needed.

If a topology is created using builder, we don't currently allow users
to edit the config with the UI text editor. They must instead import it
back into builder and edit it there. We do this ensure the builder XML
and topology JSON stay in sync.

Currently this new feature lacks support for the following, which will
be addressed at some point in the future.

1. It is currently not possible to update an existing topology when
   editing it in builder. Instead, users must (for now) create a new
   topology from any edits made to an existing one. In the future, users
   will have the option to update the existing topology being edited or
   create a new one.
2. When selecting the edit option in the Configs tab for a topology
   created using builder, we simply tell the user they must import it
   into builder instead. In the future, we will instead link directly to
   builder and import the topology for the user automatically.