Optionally identify cc clients by UUID in state of health app
Added an option to ID cc clients either by hostname or by VM UUID, with
the default being by hostname.

Updated the state of health app to include the following additional
metadata option:

`hostsToUseUUIDForC2Active`

The value for this new option should be a list of topology hostnames to
use the minimega VM UUID for when determining if their cc agent is
active. Alternatively, it can be set to `all` to signify that the VM
UUID should be used for all topology nodes.