Update base minimega Docker image to pull in new required updates
The improvements to experiment files access added in commits 3e95373 and
ccdabaf depend on some new features added to minimega that have not been
merged into minimega's master branch yet. As such, this commit updates
the minimega Dockerfile to use a Docker image for minimega that includes
the new features added to minimega.