Install yarn via npm so vmdb2 doesn't get removed from JIT image
The vmdb2 deb package depends on the cmdtest package, which provides
yarn. But this is a different yarn than the NodeJS yarn, just has the
same name. So when the NodeJS yarn is installed via its deb package, it
conflicts with cmdtest, which causes vmdb2 to be removed from the JIT
image.