Add tab to access minimega console
Very similar to what's provided in miniweb, this adds a top-level tab
that will allow interacting with the minimega console. The goal is to
allow users to accomplish low-level tasks in phenix which do not
otherwise have explicit support. Currently only phenix global admins can
access the console.

The terminal will dynamically resize (in general, any use of Terminal
can be resized if the prop is set specifying the endpoint).

A new CLI argument (minimega-path) must be set for the console to
function. The minimega Docker image has been updated to make a copy of
the minimega executable available at /tmp/minimega/bin/minimega. If the
Docker Compose file provided is being used to deploy minimega and
phenix, this means users can pass the
`--minimega-path=/tmp/minimega/bin/minimega` flag to the `phenix ui`
command when starting the UI process to make the minimega console
available.