Add Ability to Mount VMs and Download/Upload Files
When interacting with a VM in a running experiment, users can now access
files within the VM directly from the UI by clicking the `mount vm`
button within the VM's informational modal.

RBAC policies are applied to limit the ability for users to mount and
unmount VMs, as well as list, download, and upload files to and from
VMs.

Note that this feature only works if the miniccc agent is active within
the VM. If it's not, the `mount vm` button will be disabled.

Also note that this feature is also currently behind a feature flag in
the phenix UI named `vm-mount` that is not enabled by default. To enable
it, pass the `--features=vm-mount` flag to the `phenix ui` command when
starting the UI process.

Additional notes and changes:

* buefy dependency was upgraded to 0.9.22
  * this enabled use of the "breadcrumbs" component in the mount modal
  * this also required updating all tables due to a breaking change
* added flag for "cc active" to the frontend which is visible in the vm
  info modal
* created common vue filter for formatting file sizes
* fixed developer jwt signing key and added comment documenting use
* generalized ExperimentFile object to File and removed some unused code
  for ExperimentFiles in protobuf
* changed ESLint rules to not warn of unused variables if the variable
  starts with _