Update minimega.py after phenix-apps install
The `_run` command in v2.7 of minimega.py currently only adds command
args to the minimega command to be run if the args are "truthy". A
specific case where this breaks things is when the scorch mm component
is trying to configure minimega to capture packets on interface `0` of a
VM. Since `0` is "falsy", the interface index doesn't get included in
the command.

This commit adds a `sed` command to the Dockerfile that updates
minimega.py's `_run` command to add command args as long as they aren't
`None`. This is being done in the phenix Dockerfile instead of somewhere
within the phenix-apps install phase since the Python version (which is
needed for the path passed to `sed`) will be consistent.

Once a new version of minimega.py is released with a fix, this commit
can (and should) be reverted.