feat: allow users to specify external nodes in topology
It may be the case that a user would want to include nodes in the
topology that don't necessarily need to be deployed by phenix in
minimega. For example, lets say a user app exists to deploy some
topology nodes in a cloud provider. Or, lets say the state of health app
needs to know about hardware-in-the-loop nodes to do custom connectivity
tests against.

phenix now supports `external` nodes that are not deployed in minimega,
but still show up in the list of nodes in a topology, still show up in
the state of health view, etc.

A node can be marked as external by adding the top-level `external: true`
setting to a node's configuration in the topology configuration file.