fix: include VM disk images only present on remote cluster nodes
This fixes a phēnix VM list web page loading error that occurs when a
disk image only exists on a remote node. This is the case when doing a
VM snapshot for a VM that exists on a remote node. The minimega `disk
info` command, which wraps `qemu-img info`, is ran on the local node and
does not appear to account for disk images on remote nodes.