fix(vrouter): support NTP configuration via ntp app
This commit updates the vrouter app to look for the ntp app in the
scenario config and configure ntp (client only) in the Vyos/Vyatta
config files that are generated by the vrouter app.

Note that the latest version of this app does not yet support NTP
client configuration for minirouter-based routers, nor does it support
configuring routers to act as NTP servers.