feat: new deploy mode option for CLI and UI
This commit replaces the previous `runLocal` setting for experiments
with a new `deployMode` setting that can be set via the CLI and the UI
when creating new experiments.

The value of this new setting must be one of `all`, `no-headnode`, or
`only-headnode`, with the default being `no-headnode` to preserve the
previous default value of `runLocal: false`.

The overall goal of this commit is to support the new option of `all`,
which allows experiment VMs to run on the head node as well as all the
other compute nodes. The previous `runLocal` setting only supported the
current `no-headnode` and `only-headnode` options.