fix(soh): location of custom test scripts now get referenced correctly
This commit reverts a change made in commit 2bff3622 that prefixed the
location of custom test scripts in VMs with the experiment namespace.

minimega cc will search for a file to be sent in nested directories, but
will not prefix it with the nested directory in the VM if the file was
not prefixed with the nested directory when provided to the cc send
command.

fixes #175