feat: user annotation improvements
This adds a few features in the UI for making notes and customizing the
way a VM appears in the SOH graph. The intent is to make it easier for
users to track/manage the status of VMs within an experiment (e.g., a
user may mark a node red and add a note "reached by red team").

* Adds a new modal for viewing/editing the labels of a VM
  * Separate "notes" section that allows multiline, maintains insertion
    order, and hides key
  * Available in experiment pages and SOH
  * Treats minimega tags and phenix labels as one (meaning persistent
    across experiment runs, and also available in mm cli)

* Changes SOH details modal to show notes at the top
  * Adds buttons to edit labels and override style
  * New style modal lets user override the fill color, stroke color, and
    stroke style of a node. This takes precedence over any default
    styling