feat(ui): add optional --unix-socket-gid flag to the ui subcommand
Adds an optional (and hidden) flag for specifying a group ID to apply
group ownership to the unix-socket. Also sets the permissions to be
group-writable.

Use case: phenix starts from within a docker container as the root user,
but normal users wish to be able to use the unix-socket to communicate
with phenix (without having to switch to root). This allows the users to
be a member of the group specified in the flag and have write access to
the socket.