fix(builder): update picture path when importing a Builder XML file
Previously, the import from disk functionality only read in the XML data
from the imported file and displayed the resulting topology to the user.

These changes first get the current stencil path (which is where all the
pictures used to create a visual topology are stored). Regex is then
used to determine if the path currently used for pictures is different
than the current stencil path. If they are different, the current
stencil path replaces the appropriate portion of the old path for
pictures. This prevents issues with the import from disk functionality
that prior to these changes could import with broken pictures.