Startup App | Ensure Auto Logon User is Local Admin
In instances where miniccc is started by an account that is not a
Local Administrator it fails. This addition ensures that the auto
logon user is added to the local machines Administrators group which
allows miniccc access to global variables.

This obviously has security implications. I am not sure what the
'right' fix is, if we can change miniccc to work with reduced
privileges, but am openning an issue there too.