fix(vrouter): ACL rules are now enforced on VyOS
We add a jump rule for each ingress/egress ruleset in the vrouter VyOS
template so that the forward chain jumps traffic to the appropriate
name chains that are already created. Note: using the forward chain
is a limitation (does not protect the fireall, requires forwarding
traffic across interfaces).

Before this fix, firewall (ACL) rules in vrouter vyos template were
created, but added to a custom 'name' chain. Custom name chains are
not used by default. "Custom firewall chains can be created, with
commands set firewall ipv4 name <name> .... In order to use such
custom chain, a rule with action jump, and the appropriate target
should be defined in a base chain."
https://docs.vyos.io/en/latest/configuration/firewall/ipv4.html

I think a re-write of the VyOS template that takes advantage of VyOS
zone-based routing would be more efficient and intuitive to users,
but would require modification of the vrouter app spec.

ref: https://github.com/sandialabs/sceptre-phenix/issues/257